collect data scanners in a single module
