+++ nsds_lab_to_nwb/nwb_builder.py
@@ -8,8 +8,7 @@ import pytz
 from pynwb import NWBHDF5IO, NWBFile
 from pynwb.file import Subject
 
-from nsds_lab_to_nwb.common.auditory_data_scanner import AuditoryDataScanner
-from nsds_lab_to_nwb.common.behavior_data_scanner import BehaviorDataScanner
+from nsds_lab_to_nwb.inputs.data_scanners import AuditoryDataScanner
 from nsds_lab_to_nwb.metadata.metadata_manager import MetadataManager
 
 from nsds_lab_to_nwb.components.device.device_originator import DeviceOriginator
